module/speech2text/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
fastapi==0.115.8
ffmpeg==1.4
filelock==3.13.1
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
greenlet==3.2.1
h11==0.14.0
huggingface-hub==0.28.1
hyperpyyaml==1.2.2
idna==3.10
intel-cmplr-lib-rt==2025.0.2
intel-cmplr-lib-ur==2025.0.2
intel-cmplr-lic-rt==2025.0.2
intel-pti==0.10.0
intel-sycl-rt==2025.0.2
jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
optuna==4.3.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
primepy==1.3
propcache==0.3.1
protobuf==6.30.2
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytorch-triton-xpu==3.2.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
sqlalchemy==2.0.40
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tcmlib==1.2.0
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.6.0+xpu
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.6.0+xpu
torchmetrics==1.7.1
torchvision==0.21.0+xpu
tqdm==4.67.1
transformers==4.48.2
typer==0.15.2
typing-extensions==4.12.2
tzdata==2025.2
umf==0.9.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.20.0
